Restructure and remove old filerepo
