+++ tests/Pest.php
@@ -1,6 +1,8 @@
 <?php
 
 use Itiden\Backup\Contracts\Repositories\BackupRepository;
+use Itiden\Backup\Repositories\FileBackupRepository;
+use Itiden\Backup\Repositories\YamlBackupRepository;
 use Itiden\Backup\Tests\TestCase;
 use Statamic\Facades\Role;
 use Statamic\Facades\User;
@@ -8,6 +10,10 @@ use Statamic\Facades\User;
 uses(TestCase::class)
     ->afterEach(fn () => app(BackupRepository::class)->empty())
     ->in(__DIR__);
+
+dataset('backup_repositories', [
+    FileBackupRepository::class
+]);
 
 function user()
 {
